Enable class data sharing for all modes
Class data sharing speeds up startup time of Java programs, which is
especially usefull for small console programs (Maven, Ant, etc.). It
requires the file classes.jsa to be generated by install script, which
was done only for client mode on i386. Generate it also in server mode
and on amd64, but note that in server mode (and amd64 has only server
mode) it must be explicitly enabled with -Xshare:on.