Fix building on lenny with lib32gcc1 from lenny-backports on amd64
On amd64 lenny, lib32gcc1 is installed in /emul/ia32-linux, lib32gcc1
from lenny-backports, however, is installed in /usr/lib32 (symlinked
to /emul/ia32-linux). This causes dpkg-shlibdeps to fail because it is
unable to find libgcc_s.so.1. The same problem is present for
lib32asound2. Workaround it by creating shlibs.local for ia32 build on
amd64.