Add `get-orig-source` target to debian/rules file
The `get-orig-source` target is used to obtain the most recent version of the
original source package from an upstream archive.

In this case we download the required archives (JDK in 32-bit and 64-bit variants,
as well as the Unlimited JCE Policy files) directly from Oracle's CDN.

This eliminates the manual step in building the Oracle Java Debian packages.